Ability to Download Stylesheets
People probably want to be able to download their stylesheet if needed.
This adds a button on the index page where they can download the
minified and unminified source files.

I actually needed to minify the files so that was added as a flag in
sassc compliation.

A routes update was required to be able to parse id's that contained 2
dots. '.min.css' would break routing and get a 404. The constraing
circumvents this by changing the regex a bit.

Also updated was the filename that is outputted when it gets downloaded.
I need to confirm that this does not break the embed.