Add Color Variable Type
I am introducing the concept of variable types to make things more
rigid. The goal is to minimize sass coding mistakes and provide easier
ways to enter data. Color is the first variable type and it allows one
to use a color picker directly from the field instead of having to type
in a color manually.